third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-1/commit-selector-mark-matching-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that matching selectors are marked properly after new rule creation and selector change.


Running: setUp

Running: addRule
[expanded] 
element.style { ()

[expanded] 
foo, [$#inspected, $].bar, [$#inspected$] { (inspector-stylesheet:1 -> inspector-stylesheet:1:36)

[expanded] 
[$div$] { (user agent stylesheet)
    display: block;


Running: changeSelector
[expanded] 
element.style { ()

[expanded] 
[$#inspected, $]a, hr { (inspector-stylesheet:1 -> inspector-stylesheet:1:20)

[expanded] 
[$div$] { (user agent stylesheet)
    display: block;
